discordbot/requirements.txt
discord.py
sqlalchemy-aio